[Silver III] Title: 주유소, Time: 356 ms, Memory: 35132 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -4,8 +4,8 @@ import java.util.*;
 public class Main {
     public long solution(int[] arr1, int[] arr2) {
         int num = arr2[0];
-        int count = 0;
-        int answer = arr2[0] * arr1[0];
+        long count = 0;
+        long answer = (long) arr2[0] * (long) arr1[0];
         for (int i = 1; i < arr1.length; i++) {
             if (arr2[i] < num) {
                 answer += num * count;
